setting up db with models, fixed some errors
