+++ pages/urls.py
@@ -6,5 +6,5 @@ urlpatterns = [
     path('', views.home, name='home'),
     path("forum/", views.forum, name="forum"),
     path("loginpage/", views.loginpage, name="loginpage"),
-    path('search-suggestions/', views.search_suggestions, name = 'search_suggestions'),
+    path('search-suggestions/', views.character_name_suggestions, name = 'search_suggestions'),
 ]
